pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "tts-viscot"
version = "0.1.0"
description = "Test-Time Scaling for Visual Chain-of-Thought Reasoning"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "transformers>=4.40.0",
    "accelerate>=0.29.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "datasets>=2.18.0",
    "huggingface-hub>=0.22.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "pydantic>=2.0.0",
    "opencv-python>=4.8.0",
    "albumentations>=1.3.0",
    "scikit-image>=0.21.0",
    "requests>=2.31.0",
    "omegaconf>=2.3.0",
    "hydra-core>=1.3.0",
    "pyyaml>=6.0.1",
    "scipy>=1.13.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "loguru>=0.7.2",
    "rich>=13.7.0",
    "jsonlines>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-cov>=5.0.0",
    "black>=24.3.0",
    "ruff>=0.3.0",
]
tracking = [
    "wandb>=0.16.0",
    "plotly>=5.20.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short -q --cov=src --cov-report=term-missing"
markers = ["integration: marks slow tests that require network or GPU (deselect with -m 'not integration')"]
